Set Board Permissions

+++ src/boards/boards.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { User } from 'src/users/entities/user.entity';
+import { getConnection, Repository } from 'typeorm';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { UpdateBoardDto } from './dto/update-board.dto';
 import { Board } from './entities/board.entity';
@@ -12,28 +13,51 @@ export class BoardsService {
     private boardRepository: Repository<Board>,
   ) {}
 
-  async create(createBoardDto: CreateBoardDto): Promise<Board> {
-    return await this.boardRepository.save(createBoardDto);
+  async create(createBoardDto: CreateBoardDto, user: User): Promise<Board> {
+    const board = {
+      ...createBoardDto,
+      users: [user],
+    };
+
+    return await this.boardRepository.save(board);
   }
 
-  async findAll() {
-    return await this.boardRepository.find({
+  async findAll(user: User) {
+    const boards = await getConnection()
+      .createQueryBuilder()
+      .select('board')
+      .from(Board, 'board')
+      .leftJoinAndSelect('board.users', 'users')
+      .where('users.id = :id', { id: user.id })
+      .getMany();
+    return boards;
+  }
+
+  async findOne(id: number, user: User) {
+    const board = await this.boardRepository.findOneOrFail(id, {
       relations: ['users'],
     });
+    const hasAccess = board.users.find(({ id }) => id === user.id);
+    if (!hasAccess) throw new UnauthorizedException();
+    return board;
   }
 
-  async findOne(id: number) {
-    return await this.boardRepository.findOneOrFail(id);
-  }
-
-  async update(id: number, updateBoardDto: UpdateBoardDto) {
-    await this.boardRepository.findOneOrFail(id);
-    await this.boardRepository.update(id, updateBoardDto);
+  async update(id: number, updateBoardDto: UpdateBoardDto, user: User) {
+    const board = await this.boardRepository.findOneOrFail(id, {
+      relations: ['users'],
+    });
+    const hasAccess = board.users.find(({ id }) => id === user.id);
+    if (!hasAccess) throw new UnauthorizedException();
+    await this.boardRepository.save({ id, ...updateBoardDto });
     return this.boardRepository.findOne(id);
   }
 
-  async remove(id: number) {
-    await this.boardRepository.findOneOrFail(id);
+  async remove(id: number, user: User) {
+    const board = await this.boardRepository.findOneOrFail(id, {
+      relations: ['users'],
+    });
+    const hasAccess = board.users.find(({ id }) => id === user.id);
+    if (!hasAccess) throw new UnauthorizedException();
     await this.boardRepository.delete(id);
     return;
   }
